requirements.txt
faster-whisper==1.0.2
groq==0.6.0
openai
google-generativeai==0.5.4
opencv_python==4.9.0.80
Pillow==10.3.0
PyAudio==0.2.14
SpeechRecognition==3.10.1
pyperclip==1.8.2